backend/requirements.txt
Django>3.1,<3.3
django-cleanup==6.0.0
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
ipython==8.4.0
prompt_toolkit==3.0.29
kombu==5.2.4
psycopg2-binary==2.9.3
ptyprocess==0.7.0
raven==6.10.0
Unidecode==1.3.4
django-crispy-forms==1.14.0
py3DNS==3.2.1
validate-email==1.3
django-recaptcha==3.0.0
django-tempus-dominus==5.1.2.6.post1
bleach[css]==5.0.0
factory-boy==3.2.1
django-ordered-model==3.6
django-select2==7.10.0
django-environ==0.8.1
django-summernote==0.8.20.0
Pillow==9.1.1
django-redis==5.2.0
django-cachalot==2.5.1